Change card delete functionality
Now instead of providing card delete functionality, a basic message is shown to users who hold a license containing the "See Message" capability. This change has been made to remove any focus on the specific features of the boilerplate; this will shift emphasis to the Salable implementation details which are more important.